[REFACTOR] repository 파일 패키지 구조 변경 (#519)
