+++ app/src/main/java/org/sopt/havit/data/repository/SearchRepositoryImpl.kt
@@ -1,12 +1,12 @@
-package org.sopt.havit.domain.repository
+package org.sopt.havit.data.repository
 
 import android.content.Context
 import org.sopt.havit.data.RetrofitObject
 import org.sopt.havit.data.remote.ContentsSearchResponse
-import org.sopt.havit.data.repository.SearchRepository
+import org.sopt.havit.domain.repository.SearchRepository
 import org.sopt.havit.util.MySharedPreference
 
-class SearchRepositoryImpl(context: Context) :SearchRepository {
+class SearchRepositoryImpl(context: Context) : SearchRepository {
 
     private val pref = MySharedPreference.getXAuthToken(context)
